fix: store data not propagating
There is a quite complex bug with the browserStore class. If updated rapidly it stop writing on the
store and the storage. So it seems to be a race condition bug. I tried various approach but the
only one working for me is to check if storage write is finish